pyproject.toml
[project]
name = "security-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "binarycraft007", email = "elliot.huang.signed@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "hvac>=2.4.0",
    "ping3>=5.1.5",
    "prometheus-client>=0.23.1",
    "psutil>=7.1.3",
    "pyjwt>=2.10.1",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
